# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Error Correction / 智卷系统 Backend API",
    "version": "1.0.0",
    "description": "基于当前 Flask 后端源码整理的 OpenAPI/Swagger 文档，供 App 端开发对接。认证采用 Flask session cookie；App 端需保留并随请求携带登录后下发的 Cookie。"
  },
  "servers": [
    {
      "url": "http://localhost:5001",
      "description": "本地后端服务"
    },
    {
      "url": "/",
      "description": "同源或前端代理根路径"
    }
  ],
...
  "security": [
    {
      "SessionCookie": []
    }
  ],
  "paths": {
...
    "/api/device/bind": {
      "post": {
        "tags": [
          "Device"
        ],
        "summary": "创建或获取当前用户的硬件设备 UUID",
        "description": "登录用户调用。后端生成并持久化 device_uuid -> user_id 绑定关系；App/手机端随后通过蓝牙把 device_uuid 写入硬件。",
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "force_new": {
                    "type": "boolean",
                    "description": "???????? device_uuid??? false????????????",
                    "default": false
                  }
                },
                "additionalProperties": false
              },
              "example": {
                "force_new": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "??",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "device_uuid",
                    "qr_payload"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "device_uuid": {
                      "type": "string",
                      "format": "uuid",
                      "example": "8f4b8f6e-2c7a-4e3a-9c6a-4c1f2e7b9a21"
                    },
                    "qr_payload": {
                      "type": "string",
                      "example": "aiwb://bind?device_uuid=8f4b8f6e-2c7a-4e3a-9c6a-4c1f2e7b9a21"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "???",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/device/capture": {
      "post": {
        "tags": [
          "Device"
        ],
        "summary": "硬件设备上传 JPEG 拍照结果",
        "description": "固件调用，无需登录 Cookie。multipart/form-data 字段：device_uuid=<后端分配并由 App 蓝牙写入硬件的 UUID>，image=<JPEG 文件>。兼容旧字段名 device_id。",
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "device_uuid",
                  "image"
                ],
                "properties": {
                  "device_uuid": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "device_id": {
                    "type": "string",
                    "description": "??????????? device_uuid?"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "JPEG ??????????? image?"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "??????????? HTTP 2xx ????",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "device_uuid",
                    "capture_id",
                    "file_key",
                    "filename"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "capture uploaded"
                    },
                    "device_uuid": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "capture_id": {
                      "type": "integer"
                    },
                    "file_key": {
                      "type": "string"
                    },
                    "filename": {
                      "type": "string",
                      "example": "capture.jpg"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "??????????",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "?????",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "????????",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/device/unbind": {
      "post": {
        "tags": [
          "Device"
        ],
        "summary": "解绑当前用户的硬件设备 UUID",
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "device_uuid"
                ],
                "properties": {
                  "device_uuid": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "device_id": {
                    "type": "string",
                    "description": "??????????? device_uuid?"
                  }
                },
                "additionalProperties": false
              },
              "example": {
                "device_uuid": "8f4b8f6e-2c7a-4e3a-9c6a-4c1f2e7b9a21"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "??",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "description": "????",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "?????????????",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/device/images": {
      "get": {
        "tags": [
          "Device"
        ],
        "summary": "查询当前用户的硬件上传图片列表",
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "device_uuid",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "????????????????????????????????"
          },
          {
            "name": "device_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "??????????? device_uuid?"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "minimum": 1,
              "maximum": 200
            }
          }
        ],
        "responses": {
          "200": {
            "description": "??",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "images"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "images": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/DeviceCapture"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "?????????????",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/device/images/use": {
      "post": {
        "tags": [
          "Device"
        ],
        "summary": "????????? AI ???",
        "description": "???????? /api/device/images ???? capture_id ???? AI ??????????????? /api/ocr ? /api/split ?????????????",
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "capture_ids"
                ],
                "properties": {
                  "capture_ids": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    },
                    "description": "?????? ID ???"
                  },
                  "reset_session": {
                    "type": "boolean",
                    "default": true,
                    "description": "????????????????? true?"
                  }
                },
                "additionalProperties": false
              },
              "example": {
                "capture_ids": [
                  1
                ],
                "reset_session": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "??",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "result"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string"
                    },
                    "result": {
                      "type": "object",
                      "properties": {
                        "file_count": {
                          "type": "integer"
                        },
                        "files": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "capture_id": {
                                "type": "integer"
                              },
                              "file_key": {
                                "type": "string"
                              },
                              "filename": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "????",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "???????????????",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "410": {
            "description": "?????????",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Flask session cookie。登录/注册成功后由后端 Set-Cookie 写入；App 端需持久化并在后续请求携带。"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        },
        "additionalProperties": true
      },
      "OkResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string"
          }
        },
        "additionalProperties": true
      },
...
      "DeviceCapture": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "device_uuid": {
            "type": "string",
            "format": "uuid"
          },
          "file_key": {
            "type": "string"
          },
          "filename": {
            "type": "string"
          },
          "image_url": {
            "type": "string",
            "example": "/api/image/abc123.jpg"
          },
          "content_type": {
            "type": "string",
            "example": "image/jpeg"
          },
          "file_size": {
            "type": "integer"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": true
      }
    }
  }
}